Move MultiBind helpers + tests to separate files wip Complete MultiBind unit tests
