+++ Xamarin.Forms.Core.UnitTests/Markup/BindableObjectExtensionsTests.cs
@@ -1,13 +1,9 @@
 ﻿using System;
-using System.Reflection;
 using NUnit.Framework;
 
 namespace Xamarin.Forms.Markup.UnitTests
 {
-	using System.Collections.Generic;
-	using System.Globalization;
 	using System.Linq;
-	using System.Net.Http.Headers;
 	using System.Windows.Input;
 	using XamarinFormsMarkupUnitTestsBindableObjectViews;
 
@@ -583,184 +579,6 @@ namespace Xamarin.Forms.Markup.UnitTests
 			);
 		});
 
-
-		// TODO: Complete unit tests for multibind
-
-		[Test]
-		[TestCase(true, false)]
-		[TestCase(false, true)]
-		[TestCase(true, true)]
-		public void BindSpecifiedPropertyWith4BindingsAndInlineConvert(bool testConvert, bool testConvertBack)
-		{
-			var label = new Label();
-
-			var bindings = new List<BindingBase> {
-				new Binding(nameof(viewModel.Text)),
-				new Binding(nameof(viewModel.Id)),
-				new Binding(nameof(viewModel.IsRed)),
-				new Binding(nameof(viewModel.Fraction))
-			};
-
-			// Repeat inline converter code for each variantion to test that the Bind overloads allow inferring the generic conversion parameter types
-			// The developer should not have to specify generic type parameters
-			if (testConvert && testConvertBack)
-			{
-				label.Bind(
-					Label.TextProperty,
-					bindings[0], bindings[1], bindings[2], bindings[3],
-					((string text, Guid id, bool isRed, double fraction) values) => $"Text = '{ values.text }', Id = '{ values.id }', IsRed = '{ values.isRed }', Fraction = '{ values.fraction }'",
-					(string formatted) =>
-					{
-						var split = formatted.Split('\'');
-						return (
-							split[1],
-							Guid.Parse(split[3]),
-							bool.Parse(split[5]),
-							double.Parse(split[7])
-						);
-					}
-				);
-			} else if (testConvert && !testConvertBack) {
-				label.Bind(
-					Label.TextProperty,
-					bindings[0], bindings[1], bindings[2], bindings[3],
-					((string text, Guid id, bool isRed, double fraction) values) => $"Text = '{ values.text }', Id = '{ values.id }', IsRed = '{ values.isRed }', Fraction = '{ values.fraction }'"
-				);
-			}
-			else if (!testConvert && testConvertBack) {
-				label.Bind(
-					Label.TextProperty,
-					bindings[0], bindings[1], bindings[2], bindings[3],
-					convertBack: (string formatted) =>
-					{
-						var split = formatted.Split('\'');
-						return (
-							split[1],
-							Guid.Parse(split[3]),
-							bool.Parse(split[5]),
-							double.Parse(split[7])
-						);
-					}
-				);
-			}
-
-			var avalues = new object[] {
-				"Hi",
-				Guid.Parse("{272383A4-92E3-46BA-99DC-438D81E039AB}"),
-				true,
-				0.5
-			};
-
-			BindingHelpers.AssertBindingExists<string>(
-				label,
-				targetProperty: Label.TextProperty,
-				bindings,
-				assertConverterInstanceIsAnyNotNull: true,
-				assertConvert: c => { c.AssertConvert(
-					avalues, 
-					$"Text = '{ avalues[0] }', Id = '{ avalues[1] }', IsRed = '{ avalues[2] }', Fraction = '{ avalues[3] }'",
-					twoWay: testConvert && testConvertBack,
-					backOnly: !testConvert && testConvertBack
-				); }
-			);
-		}
-
-		[Test]
-		[TestCase(true, false)]
-		[TestCase(false, true)]
-		[TestCase(true, true)]
-		public void BindSpecifiedPropertyWithMultipleBindings(bool testConvert, bool testConvertBack)
-		{
-			var label = new Label();
-			var bindings = new List<BindingBase> {
-				new Binding(nameof(viewModel.Text)),
-				new Binding(nameof(viewModel.Id)),
-				new Binding(nameof(viewModel.IsRed)),
-				new Binding(nameof(viewModel.Fraction)),
-				new Binding(nameof(viewModel.Count)),
-			};
-
-			Func<object[], string> convert = null;
-			if (testConvert) convert = (object[] values) => $"Text = '{ values[0] }', Id = '{ values[1] }', IsRed = '{ values[2] }', Fraction = '{ values[3] }', Count = '{ values[4] }'";
-
-			Func<string, object[]> convertBack = null;
-			if (testConvertBack)
-				convertBack = (string text) => {
-					var split = text.Split('\'');
-					return new object[]
-					{
-						split[1],
-						Guid.Parse(split[3]),
-						bool.Parse(split[5]),
-						double.Parse(split[7]),
-						int.Parse(split[9])
-					};
-				};
-
-			var converter = new FuncMultiConverter<string, object>(convert, convertBack);
-
-			label.Bind(
-				Label.TextProperty,
-				bindings,
-				converter
-			);
-
-			var avalues = new object[] {
-				"Hi",
-				Guid.Parse("{272383A4-92E3-46BA-99DC-438D81E039AB}"),
-				true,
-				0.5,
-				3
-			};
-
-			BindingHelpers.AssertBindingExists<string>(
-				label,
-				targetProperty: Label.TextProperty,
-				bindings,
-				converter,
-				assertConvert: c => { c.AssertConvert(
-					avalues, 
-					$"Text = '{ avalues[0] }', Id = '{ avalues[1] }', IsRed = '{ avalues[2] }', Fraction = '{ avalues[3] }', Count = '{ avalues[4] }'",
-					twoWay: testConvert && testConvertBack,
-					backOnly: !testConvert && testConvertBack
-				); }
-			);
-		}
-
-		List<BindingBase> GetTestBindings(int count) => new List<BindingBase> {
-			new Binding(nameof(viewModel.Text)),
-			new Binding(nameof(viewModel.Id)),
-			new Binding(nameof(viewModel.IsRed)),
-			new Binding(nameof(viewModel.Fraction)),
-			new Binding(nameof(viewModel.Count))
-		}.Take(count).ToList();
-
-		object[] GetTestConvertValues(int count) => new List<object> {
-			"Hi",
-			Guid.Parse("{272383A4-92E3-46BA-99DC-438D81E039AB}"),
-			true,
-			0.5,
-			3
-		}.Take(count).ToArray();
-
-		void AssertLabelTextMultiBound(Label label, int nBindings, bool testConvert, bool testConvertBack, IMultiValueConverter converter = null)
-		{
-			var values = GetTestConvertValues(nBindings);
-			string expected               = $"Text = '{ values[0] }', Id = '{ values[1] }'";
-			if (nBindings >= 3) expected += $", IsRed = '{ values[2] }'";
-			if (nBindings >= 4) expected += $", Fraction = '{ values[3] }'";
-			if (nBindings >= 5) expected += $", Count = '{ values[4] }'";
-
-			BindingHelpers.AssertBindingExists<string>(
-				label,
-				targetProperty: Label.TextProperty,
-				GetTestBindings(nBindings),
-				converter,
-				assertConverterInstanceIsAnyNotNull: converter == null,
-				assertConvert: c => c.AssertConvert(values, expected, twoWay: testConvert && testConvertBack, backOnly: !testConvert && testConvertBack)
-			);
-		}
-
 		[Test]
 		public void BindCommandWithDefaults() => AssertExperimental(() =>
 		{
@@ -852,180 +670,7 @@ namespace Xamarin.Forms.Markup.UnitTests
 			public string Text { get; set; }
 			public Color TextColor { get; set; }
 			public bool IsRed { get; set; }
-			public double Fraction { get; set; }
-			public int Count { get; set; }
 		}
-	}
-
-	internal static class BindingHelpers
-	{
-		static MethodInfo getContextMethodInfo;
-		static FieldInfo bindingFieldInfo;
-
-		internal static void AssertBindingExists(
-			BindableObject bindable,
-			BindableProperty targetProperty,
-			string path = ".",
-			BindingMode mode = BindingMode.Default,
-			bool assertConverterInstanceIsAnyNotNull = false,
-			IValueConverter converter = null,
-			object converterParameter = null,
-			string stringFormat = null,
-			object source = null,
-			object targetNullValue = default,
-			object fallbackValue = default,
-			Action<IValueConverter> assertConvert = null
-		) => AssertBindingExists<object, object>(
-				bindable, targetProperty, path, mode, assertConverterInstanceIsAnyNotNull, converter, converterParameter,
-				stringFormat, source, targetNullValue, fallbackValue, assertConvert
-		);
-
-		internal static void AssertBindingExists<TDest>(
-			BindableObject bindable,
-			BindableProperty targetProperty,
-			string path = ".",
-			BindingMode mode = BindingMode.Default,
-			bool assertConverterInstanceIsAnyNotNull = false,
-			IValueConverter converter = null,
-			string stringFormat = null,
-			object source = null,
-			TDest targetNullValue = default,
-			TDest fallbackValue = default,
-			Action<IValueConverter> assertConvert = null
-		) => AssertBindingExists<TDest, object>(
-				bindable, targetProperty, path, mode, assertConverterInstanceIsAnyNotNull, converter, null,
-				stringFormat, source, targetNullValue, fallbackValue, assertConvert
-		);
-
-		internal static void AssertBindingExists<TDest, TParam>(
-			BindableObject bindable,
-			BindableProperty targetProperty,
-			string path = ".",
-			BindingMode mode = BindingMode.Default,
-			bool assertConverterInstanceIsAnyNotNull = false,
-			IValueConverter converter = null,
-			TParam converterParameter = default,
-			string stringFormat = null,
-			object source = null,
-			TDest targetNullValue = default,
-			TDest fallbackValue = default,
-			Action<IValueConverter> assertConvert = null
-		)
-		{
-			var binding = BindingHelpers.GetBinding(bindable, targetProperty);
-			Assert.That(binding, Is.Not.Null);
-			Assert.That(binding.Path, Is.EqualTo(path));
-			Assert.That(binding.Mode, Is.EqualTo(mode));
-			if (assertConverterInstanceIsAnyNotNull)
-				Assert.That(binding.Converter, Is.Not.Null);
-			else
-				Assert.That(binding.Converter, Is.EqualTo(converter));
-			Assert.That(binding.ConverterParameter, Is.EqualTo(converterParameter));
-			Assert.That(binding.StringFormat, Is.EqualTo(stringFormat));
-			Assert.That(binding.Source, Is.EqualTo(source));
-			Assert.That(binding.TargetNullValue, Is.EqualTo(targetNullValue));
-			Assert.That(binding.FallbackValue, Is.EqualTo(fallbackValue));
-
-			assertConvert?.Invoke(binding.Converter);
-		}
-
-		internal static void AssertBindingExists<TDest>(
-			BindableObject bindable,
-			BindableProperty targetProperty,
-			IList<BindingBase> bindings,
-			IMultiValueConverter converter = null,
-			BindingMode mode = BindingMode.Default,
-			bool assertConverterInstanceIsAnyNotNull = false,
-			string stringFormat = null,
-			TDest targetNullValue = default,
-			TDest fallbackValue = default,
-			Action<IMultiValueConverter> assertConvert = null
-		) => AssertBindingExists<TDest, object>(
-			bindable, targetProperty, bindings, converter, null, mode, assertConverterInstanceIsAnyNotNull,
-			stringFormat, targetNullValue, fallbackValue, assertConvert);
-
-
-		internal static void AssertBindingExists<TDest, TParam>(
-			BindableObject bindable,
-			BindableProperty targetProperty,
-			IList<BindingBase> bindings,
-			IMultiValueConverter converter = null,
-			TParam converterParameter = default,
-			BindingMode mode = BindingMode.Default,
-			bool assertConverterInstanceIsAnyNotNull = false,
-			string stringFormat = null,
-			TDest targetNullValue = default,
-			TDest fallbackValue = default,
-			Action<IMultiValueConverter> assertConvert = null
-		)
-		{
-			var binding = BindingHelpers.GetMultiBinding(bindable, targetProperty);
-			Assert.That(binding, Is.Not.Null);
-			Assert.That(binding.Bindings.SequenceEqual(bindings), Is.True);
-			Assert.That(binding.Mode, Is.EqualTo(mode));
-			if (assertConverterInstanceIsAnyNotNull)
-				Assert.That(binding.Converter, Is.Not.Null);
-			else
-				Assert.That(binding.Converter, Is.EqualTo(converter));
-			Assert.That(binding.ConverterParameter, Is.EqualTo(converterParameter));
-			Assert.That(binding.StringFormat, Is.EqualTo(stringFormat));
-			Assert.That(binding.TargetNullValue, Is.EqualTo(targetNullValue));
-			Assert.That(binding.FallbackValue, Is.EqualTo(fallbackValue));
-
-			assertConvert?.Invoke(binding.Converter);
-		}
-
-		internal static Binding GetBinding(BindableObject bindable, BindableProperty property) => GetBindingBase(bindable, property) as Binding;
-
-		internal static MultiBinding GetMultiBinding(BindableObject bindable, BindableProperty property) => GetBindingBase(bindable, property) as MultiBinding;
-
-		/// <remarks>
-		/// Note that we are only testing whether the Markup helpers create the correct bindings,
-		/// we are not testing the binding mechanism itself; this is why it is justified to access
-		/// private binding API's here for testing.
-		/// </remarks>
-		internal static BindingBase GetBindingBase(BindableObject bindable, BindableProperty property)
-		{
-			// return bindable.GetContext(property)?.Binding as Binding;
-			// Both BindableObject.GetContext and BindableObject.BindablePropertyContext are private; 
-			// use reflection instead of above line.
-
-			if (getContextMethodInfo == null)
-				getContextMethodInfo = typeof(BindableObject).GetMethod("GetContext", BindingFlags.NonPublic | BindingFlags.Instance);
-
-			var context = getContextMethodInfo?.Invoke(bindable, new object[] { property });
-			if (context == null)
-				return null;
-
-			if (bindingFieldInfo == null)
-				bindingFieldInfo = context?.GetType().GetField("Binding");
-
-			return bindingFieldInfo?.GetValue(context) as BindingBase;
-		}
-
-		internal static IValueConverter AssertConvert<TValue, TConvertedValue>(this IValueConverter converter, TValue value, object parameter, TConvertedValue expectedConvertedValue, bool twoWay = false, bool backOnly = false, CultureInfo culture = null)
-		{
-			Assert.That(converter?.Convert(value, typeof(object), parameter, culture), Is.EqualTo(backOnly ? default(TConvertedValue) : expectedConvertedValue));
-			Assert.That(converter?.ConvertBack(expectedConvertedValue, typeof(object), parameter, culture), Is.EqualTo(twoWay || backOnly ? value : default(TValue)));
-			return converter;
-		}
-
-		internal static IValueConverter AssertConvert<TValue, TConvertedValue>(this IValueConverter converter, TValue value, TConvertedValue expectedConvertedValue, bool twoWay = false, bool backOnly = false, CultureInfo culture = null)
-			=> AssertConvert(converter, value, null, expectedConvertedValue, twoWay: twoWay, backOnly: backOnly, culture: culture);
-
-		internal static IMultiValueConverter AssertConvert<TConvertedValue>(this IMultiValueConverter converter, object[] values, object parameter, TConvertedValue expectedConvertedValue, bool twoWay = false, bool backOnly = false, CultureInfo culture = null)
-		{
-			Assert.That(converter?.Convert(values, typeof(TConvertedValue), parameter, culture), Is.EqualTo(backOnly ? BindableProperty.UnsetValue : expectedConvertedValue));
-			var convertedBackValues = converter?.ConvertBack(expectedConvertedValue, null, parameter, culture);
-			if (twoWay || backOnly)
-				Assert.That(convertedBackValues.SequenceEqual(values), Is.True);
-			else
-				Assert.That(convertedBackValues, Is.Null);
-			return converter;
-		}
-
-		internal static IMultiValueConverter AssertConvert<TConvertedValue>(this IMultiValueConverter converter, object[] values, TConvertedValue expectedConvertedValue, bool twoWay = false, bool backOnly = false, CultureInfo culture = null)
-		=> AssertConvert<TConvertedValue>(converter, values, null, expectedConvertedValue, twoWay, backOnly, culture);
 	}
 }
