+++ Xamarin.Forms.Core.UnitTests/Xamarin.Forms.Core.UnitTests.csproj
@@ -77,6 +77,8 @@
     <Compile Include="CommandTests.cs" />
     <Compile Include="DependencyResolutionTests.cs" />
     <Compile Include="ItemsLayoutTypeConverterTests.cs" />
+    <Compile Include="Markup\BindableObjectMultiBindExtensionsTests.cs" />
+    <Compile Include="Markup\BindingHelpers.cs" />
     <Compile Include="Markup\RelativeLayoutExtensionsTests.cs" />
     <Compile Include="Markup\BindableLayoutExtensionsTests.cs" />
     <Compile Include="MultiBindingTests.cs" />
